requirement.txt
# Core backend framework
fastapi==0.110.0
uvicorn[standard]==0.29.0

# Database & ORM
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
alembic==1.13.1

# Data validation
pydantic==2.6.4
pydantic-settings==2.2.1

# AI / LLM
google-generativeai==0.3.2


# Async & HTTP
httpx==0.27.0
aiofiles==23.2.1

# Utilities
python-dotenv==1.0.1
loguru==0.7.2
numpy==1.26.3
jira==3.5.2

# Testing
pytest==8.1.1
pytest-asyncio==0.23.5
